Revert "Fixing lot of win64 type conversion warnings (in progress)"
This reverts commit 8227befb55309dbca859809c29d5676efdc8f89e [formerly 4ee5fa10fd43bc451242bba96b358f70536bdc46].


Former-commit-id: f964f88841c4d6b9d52af634d69fc16194366405

+++ framework/sofa/core/objectmodel/BaseLink.cpp
@@ -66,7 +66,7 @@ void BaseLink::printValue( std::ostream& o ) const
     bool first = true;
 	for (std::size_t i = 0; i<size; ++i)
     {
-        std::string path = getLinkedPath((unsigned int)i);
+        std::string path = getLinkedPath(i);
         if (path.empty()) continue;
         if (first) first = false;
         else o << ' ';
