+++ modules/SofaUserInteraction/SleepController.cpp
@@ -39,7 +39,7 @@ bool StateTester<DataTypes>::wantsToSleep(core::behavior::BaseMechanicalState* b
 
 	const VecDeriv& vecVel = state->read(core::ConstVecDerivId::velocity())->getValue();
 	SReal maxVelocity = 0;
-	for (unsigned int i = 0, nb = (unsigned int)vecVel.size(); i < nb; ++i)
+	for (unsigned int i = 0, nb = vecVel.size(); i < nb; ++i)
 		maxVelocity = std::max(maxVelocity, vecVel[i].norm2());
 
 	return maxVelocity < (speedThreshold * speedThreshold);
@@ -55,7 +55,7 @@ bool StateTester<defaulttype::Rigid2Types>::wantsToSleep(core::behavior::BaseMec
 
 	const DataTypes::VecDeriv& vecVel = state->read(core::ConstVecDerivId::velocity())->getValue();
 	SReal maxSpeed = 0, maxRotation = 0;
-	for (unsigned int i = 0, nb = (unsigned int)vecVel.size(); i < nb; ++i)
+	for (unsigned int i = 0, nb = vecVel.size(); i < nb; ++i)
 	{
 		SReal speed = vecVel[i].getVCenter().norm2();
 		maxSpeed = std::max(maxSpeed, speed);
@@ -78,7 +78,7 @@ bool StateTester<defaulttype::Rigid3Types>::wantsToSleep(core::behavior::BaseMec
 
 	const DataTypes::VecDeriv& vecVel = state->read(core::ConstVecDerivId::velocity())->getValue();
 	SReal maxSpeed = 0, maxRotation = 0;
-	for (unsigned int i = 0, nb = (unsigned int)vecVel.size(); i < nb; ++i)
+	for (unsigned int i = 0, nb = vecVel.size(); i < nb; ++i)
 	{
 		SReal speed = vecVel[i].getVCenter().norm2();
 		maxSpeed = std::max(maxSpeed, speed);
@@ -123,11 +123,11 @@ void SleepController::init()
 	GetStatesThatCanSleep(core::ExecParams::defaultInstance(), tempStates).execute(getContext()->getRootContext());
 
 	// Find the corresponding template in each mechanical state we are monitoring
-	for (unsigned int i = 0, nbStates = (unsigned int)tempStates.size(); i < nbStates; ++i)
+	for (unsigned int i = 0, nbStates = tempStates.size(); i < nbStates; ++i)
 	{
 		core::behavior::BaseMechanicalState* state = tempStates[i];
 		bool found = false;
-		for (unsigned int j = 0, nbTesters = (unsigned int)m_stateTesters.size(); j < nbTesters; ++j)
+		for (unsigned int j = 0, nbTesters = m_stateTesters.size(); j < nbTesters; ++j)
 		{
 			StateTesterPtr tester = m_stateTesters[j];
 			if (tester->canConvert(state))
@@ -149,7 +149,7 @@ void SleepController::init()
 	// If it is the first time doing the init, copy the initial state values
 	if (m_initialState.empty())
 	{
-		for (unsigned int i = 0, nbContexts = (unsigned int)m_contextsThatCanSleep.size(); i < nbContexts; ++i)
+		for (unsigned int i = 0, nbContexts = m_contextsThatCanSleep.size(); i < nbContexts; ++i)
 			m_initialState.push_back(m_contextsThatCanSleep[i]->isSleeping());
 	}
 
@@ -161,12 +161,12 @@ void SleepController::reset()
 	// Reset the states of the nodes to the initial values
 	if (m_initialState.size() == m_contextsThatCanSleep.size())
 	{
-		for (unsigned int i = 0, nbContexts = (unsigned int)m_contextsThatCanSleep.size(); i < nbContexts; ++i)
+		for (unsigned int i = 0, nbContexts = m_contextsThatCanSleep.size(); i < nbContexts; ++i)
 			m_contextsThatCanSleep[i]->setSleeping(m_initialState[i]);
 	}
 
 	// Reset time since wake up to 0
-	for (unsigned int i = 0, nb = (unsigned int)m_timeSinceWakeUp.size(); i < nb; ++i)
+	for (unsigned int i = 0, nb = m_timeSinceWakeUp.size(); i < nb; ++i)
 			m_timeSinceWakeUp[i] = 0.0;
 }
 
@@ -190,7 +190,7 @@ void SleepController::putNodesToSleep()
 	
 	SReal speedThreshold = d_speedThreshold.getValue();
 	SReal rotThreshold = d_rotationThreshold.getValue();
-	for (unsigned int i = 0, nbStates = (unsigned int)m_statesThatCanSleep.size(); i < nbStates; ++i)
+	for (unsigned int i = 0, nbStates = m_statesThatCanSleep.size(); i < nbStates; ++i)
 	{
 		core::behavior::BaseMechanicalState* state = m_statesThatCanSleep[i];
 		core::objectmodel::BaseContext* context = m_contextsThatCanSleep[i];
@@ -233,7 +233,7 @@ void SleepController::wakeUpNodes()
 	do
 	{
 		changed = false;
-		unsigned int nbContexts = (unsigned int)m_contextsThatCanSleep.size();
+		unsigned int nbContexts = m_contextsThatCanSleep.size();
 		for (unsigned int i = 0; i < nbContexts; ++i)
 		{
 			core::objectmodel::BaseContext* context = m_contextsThatCanSleep[i];
@@ -241,7 +241,7 @@ void SleepController::wakeUpNodes()
 				continue;
 
 			const BaseContexts& wakeupPair = wakeupPairs[i];
-			for (unsigned int j = 0, nbLinks = (unsigned int)wakeupPair.size(); j < nbLinks; ++j)
+			for (unsigned int j = 0, nbLinks = wakeupPair.size(); j < nbLinks; ++j)
 			{
 				if (!wakeupPair[j]->isSleeping())
 				{
@@ -260,7 +260,7 @@ void SleepController::wakeUpNodes()
 
 void SleepController::updateTimeSinceWakeUp()
 {
-	for (unsigned int i = 0, nbContexts = (unsigned int)m_contextsThatCanSleep.size(); i < nbContexts; ++i)
+	for (unsigned int i = 0, nbContexts = m_contextsThatCanSleep.size(); i < nbContexts; ++i)
 	{
 		core::objectmodel::BaseContext* context = m_contextsThatCanSleep[i];
 		if (context->isSleeping())
@@ -287,7 +287,7 @@ void SleepController::collectWakeupPairs(std::vector<BaseContexts>& wakeupPairs)
 	else
 		root->get<core::collision::Contact>(&contacts, core::objectmodel::BaseContext::SearchDown);
 
-	for (unsigned int i = 0, nbContacts = (unsigned int)contacts.size(); i < nbContacts; ++i)
+	for (unsigned int i = 0, nbContacts = contacts.size(); i < nbContacts; ++i)
 	{
 		std::pair<core::CollisionModel*, core::CollisionModel*> collisionModels = contacts[i]->getCollisionModels();
 
@@ -317,7 +317,7 @@ void SleepController::addWakeupPair(std::vector<BaseContexts>& wakeupPairs, core
 		BaseContexts::const_iterator iter = std::find(contextsBegin, contextsEnd, context1);
 		if (iter != contextsEnd)
 		{
-			int index = (int)(iter - contextsBegin);
+			int index = iter - contextsBegin;
 			BaseContexts& wakeupPairRef = wakeupPairs[index];
 			if(std::find(wakeupPairRef.begin(), wakeupPairRef.end(), context2) == wakeupPairRef.end()) // No duplicates
 				wakeupPairRef.push_back(context2);
@@ -329,7 +329,7 @@ void SleepController::addWakeupPair(std::vector<BaseContexts>& wakeupPairs, core
 		BaseContexts::const_iterator iter = std::find(contextsBegin, contextsEnd, context2);
 		if (iter != contextsEnd)
 		{
-			int index = (int)(iter - contextsBegin);
+			int index = iter - contextsBegin;
 			BaseContexts& wakeupPairRef = wakeupPairs[index];
 			if(std::find(wakeupPairRef.begin(), wakeupPairRef.end(), context1) == wakeupPairRef.end()) // No duplicates
 				wakeupPairRef.push_back(context1);
